refactor(game-engine): extract handleBlitz to blitz-handler.ts (S27.8.13)
Extract the largest remaining handler from `actions.ts` (~226 lines)
into `actions/blitz-handler.ts` (282 lines). The full flow is preserved:
defensive Foul Appearance check, movement with optional dodge
(Shadowing + Break Tackle BB3 + armor + injury on failure + bounce
ball), and post-move block (assists, dice count, chooser, pendingBlock).

No cycle back to `actions.ts`: only leaf helpers from `mechanics/*`,
`core/*`, `utils/*`, `skills/*`.

Massive orphan import cleanup in `actions.ts` (25 imports removed):
`Position`, `getDodgeSkillModifiers`, `requiresDodgeRoll`,
`calculateDodgeModifiers`, `rollBlockDiceManyWithRolls`,
`performDodgeRollWithNotification`, `performArmorRollWithNotification`,
`performInjuryRoll`, `createLogEntry`, `isInOpponentEndzone`,
`awardTouchdown`, `bounceBall`, `canBlock`, `canBlitz`,
`calculateOffensiveAssists`, `calculateDefensiveAssists`,
`calculateBlockDiceCount`, `getBlockDiceChooser`, `setPlayerAction`,
`checkPlayerTurnEnd`, `handlePlayerSwitch`, `incrementTeamBlitzCount`,
`checkBreakTackle`, `resolveShadowingAfterDodge`, `checkFoulAppearance`.

`actions.ts` shrinks 626 -> 324 lines (-302 net lines, cumulative
-2382 vs origin 2706 = **~88% reduction**).

**DoD `actions.ts <= 600 lines` ACHIEVED**.

5 new smoke tests in `blitz-handler.test.ts` cover the exported API,
gates (missing attacker / missing target / illegal blitz on teammate),
and state immutability.

Coverage exclusion added in `vitest.config.ts` (same v8 source-map
quirk as all other S27.8.x extractions).

- 4923/4923 game-engine tests pass (+5 new smoke tests)
- 1787/1787 server tests pass (no regression)
- typecheck + lint + build clean (0 warnings on modified files)
- coverage 25.33% functions/branches (>= 25% threshold)

`match.ts` and `team.ts` are untouched in this slice (deferrable per
the original S27.8 scope note).
