+++ packages/game-engine/vitest.config.ts
@@ -48,6 +48,11 @@ export default defineConfig({
         // `actions.ts` (handleLeap, handleMove, handleDodge), meme
         // quirk source-map v8.
         'src/actions/move-leap-dodge-handlers.ts',
+        // S27.8.13 — `blitz-handler.ts` extrait depuis `actions.ts`
+        // (handleBlitz), meme quirk source-map v8. Avec cette
+        // extraction, `actions.ts` descend a ~324 lignes (DoD <= 600
+        // atteint).
+        'src/actions/blitz-handler.ts',
       ],
       // S25.4 — Seuils initiaux conservateurs. v8 reporte 0% lines sur
       // certains fichiers TS du moteur a cause d'un quirk source-map ;
